Move create task button to tasks section and fix alignment
The create task button consumed unnecessary space on mobile devices because
it was pushed under the primary title. Moving to the task list makes better
use of the view port.

Secondly there's now horrizontal rules under the tasks list heading and create
button and the line separating tasks extends the full width of the bounding box.